% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{construct_pf}
\alias{construct_pf}
\title{Construct penalty factors}
\usage{
construct_pf(w_int, w_ext, v_int, v_ext, type)
}
\arguments{
\item{w_int}{internal weights:
numeric vector of length p with non-negative entries}

\item{w_ext}{external weights:
numeric vector of length p with non-negative entries}

\item{v_int}{exponent or factor for internal weights:
non-negative scalar}

\item{v_ext}{exponent or factor for external weights:
non-negative scalar}

\item{type}{}
}
\description{
Uses internal and external weights (for negative and positive effects)
as well as internal and external exponents/factors for these weights
to construct penalty factors.
}
