% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_matrix}
\alias{count_matrix}
\title{Metrics for sign detection}
\usage{
count_matrix(truth, estim)
}
\arguments{
\item{truth}{n times p matrix with entries in -1, 0, 1}

\item{estim}{n times p matrix with entries in -1, 0, 1}
}
\description{
Metrics for sign detection
}
