% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_vector}
\alias{count_vector}
\title{Metrics for sign detection}
\usage{
count_vector(truth, estim)
}
\arguments{
\item{truth}{vector of length p with entries in -1, 0, 1}

\item{estim}{vector of length p with entries -1, 0, 1}
}
\description{
Metrics for sign detection
}
