% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cv.multiple}
\alias{cv.multiple}
\alias{cv.transfer}
\title{Model comparison}
\usage{
cv.multiple(
  y,
  X,
  family,
  alpha = 1,
  nfolds = 10,
  method = c("glm.separate", "glm.mgaussian", "sparselink"),
  alpha.init,
  type,
  trial = FALSE
)

cv.transfer(
  y,
  X,
  family,
  alpha = 1,
  nfolds = 10,
  method = c("glm.separate", "glm.glmtrans", "sparselink"),
  alpha.init,
  type,
  trial = FALSE
)
}
\arguments{
\item{y}{\eqn{n \times q} matrix (multi-task learning)
or list of \eqn{n_k}-dimensional vectors (transfer learning)}

\item{family}{character \code{"gaussian"} or \code{"binomial"}}

\item{alpha}{elastic net mixing parameter of final regressions,
default: 1 (lasso)}

\item{nfolds}{number of cross-validation folds}

\item{alpha.init}{elastic net mixing parameter for initial regressions,
default: 0.95 (lasso-like elastic net)}

\item{type}{character}

\item{trial}{experimental argument (to be removed):
Should exponents above 1 be used? Default: \code{FALSE}}
}
\description{
Compares predictive methods for multi-task learning (\code{cv.multiple}) or
transfer learning (\code{cv.transfer}) by \eqn{k}-fold cross-validation.
}
\examples{
1+1

}
\keyword{internal}
