% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.info}
\alias{get.info}
\title{Extract dimensionality.}
\usage{
get.info(x, y, family)
}
\arguments{
\item{x}{list of \eqn{q} matrices,
with \eqn{n_k} (samples) rows and \eqn{p} columns (features),
for \eqn{k} in \eqn{1,...,q}}

\item{y}{list of \eqn{q} vectors,
with \eqn{n_k} entries,
for \eqn{k} in \eqn{1,...,q}}
}
\value{
Returns a list with the slots
\eqn{q} (scalar, number of problems),
\eqn{n} (vector of length \eqn{q}, number of samples)
and \eqn{p} (scalar, number of features)
}
\description{
Extract dimensionality.
}
\examples{
data <- sim.data.transfer()
get.info(x=data$X_train,y=data$y_train,family=data$family)

}
