% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{glm.comb}
\alias{glm.comb}
\title{Sparse regression for related problems}
\usage{
glm.comb(x, y, family, alpha.init = 0.95, alpha = 1, nfolds = 10, type)
}
\arguments{
\item{x}{n x p matrix (multi-task learning) 
or list of n_k x p matrices (transfer learning)}

\item{y}{n x q matrix (multi-task learning)
or list of n_k-dimensional vectors (transfer learning)}

\item{family}{character "gaussian" or "binomial"}

\item{alpha.init}{elastic net mixing parameter for initial regressions,
default: 0.95 (lasso-like elastic net)}

\item{alpha}{elastic net mixing parameter of final regressions,
default: 1 (lasso)}

\item{nfolds}{number of cross-validation folds}

\item{type}{character}
}
\description{
Sparse regression for related problems
}
