% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logit}
\alias{logit}
\title{logit function}
\usage{
logit(x)
}
\arguments{
\item{x}{numeric vector with values in unit interval}
}
\value{
Returns a numeric vector of the transformed values.
}
\description{
logit function
}
\examples{
x <- seq(from=0,to=1,length.out=100)
y <- logit(x=x)
graphics::plot(x=x,y=y,type="l")
graphics::abline(v=0.5,lty=2)
graphics::abline(h=0,lty=2)

}
\keyword{internal}
