% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make.folds.trans}
\alias{make.folds.trans}
\title{Folds for transfer learning}
\usage{
make.folds.trans(y, family, nfolds = 10)
}
\arguments{
\item{y}{list of \eqn{q} numeric vectors}

\item{family}{character \code{"gaussian"} or \code{"binomial"}}

\item{nfolds}{integer between 2 and \eqn{n}}
}
\description{
Folds for transfer learning
}
\examples{
family <- "binomial"
y <- sim.data.transfer(family=family)$y_train
fold <- make.folds.trans(y,family=family)

}
\keyword{internal}
