% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotWeight}
\alias{plotWeight}
\title{Visualise metric that depends on two parameters}
\usage{
plotWeight(x, y)
}
\arguments{
\item{x}{list with slots source and target}

\item{y}{numeric vector}
}
\description{
Displays values in y in a grey scale (white=lowest, black=highest),
for different combinations of the two variables in x.
The lowest value is indicated by a red cross,
and the lowest value on the diagonal is indicated by a red circle.
}
\examples{
values <- seq(from=0,to=1,by=0.2)
x <- expand.grid(source=values,target=values)
y <- stats::rexp(n=length(values)*length(values))
plotWeight(x=x,y=y)

}
\keyword{internal}
