% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.sparselink}
\alias{print.sparselink}
\title{Print \code{sparselink} object}
\usage{
\method{print}{sparselink}(x, ...)
}
\arguments{
\item{x}{object of class \code{sparselink}
(generated by function \link{sparselink})}

\item{...}{(not applicable)}
}
\value{
Returns \code{NULL}. Writes information on the \code{sparselink}-object to the console.
}
\description{
Prints an object of class \code{sparselink}
}
\examples{
n <- 100; p <- 50; q <- 3
family <- "gaussian"
x <- matrix(data=rnorm(n=n*p),nrow=n,ncol=p)
y <- matrix(data=rnorm(n*q),nrow=n,ncol=q)
object <- sparselink(x=x,y=y,family=family)
object

}
\keyword{internal}
