% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Returns a numeric vector of the transformed values.
}
\description{
Sigmoid function
}
\examples{
x <- seq(from=-3,to=3,length.out=100)
y <- sigmoid(x)
graphics::plot(x=x,y=y,type="l")
graphics::abline(v=0,lty=2)
graphics::abline(h=0.5,lty=2)

}
\keyword{internal}
