% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim.data.multiple}
\alias{sim.data.multiple}
\title{Data simulation for multi-task learning}
\usage{
sim.data.multiple(
  prob.common = 0.05,
  prob.separate = 0.05,
  q = 3,
  n0 = 100,
  n1 = 10000,
  p = 200,
  family = "gaussian"
)
}
\arguments{
\item{prob.common}{probability of common effect}

\item{prob.separate}{probability of separate effect}

\item{q}{number of datasets: integer}

\item{n0}{number of training samples: integer vector of length q}

\item{n1}{number of testing samples for all datasets: integer}

\item{p}{number of features: integer}

\item{family}{character "gaussian" or "binomial"}
}
\value{
Returns list with slots
y_train (n0 x q matrix),
X_train (n0 x p matrix),
y_test (n1 x q matrix),
X_test (n1 x p matrix),
and beta (p x q matrix).
}
\description{
Simulates data for multi-task learning.
}
\examples{
data <- sim.data.multiple()
sapply(X=data,FUN=dim)
 
}
