% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{sparselink-package}
\alias{sparselink-package}
\title{Sparse regression for related problems}
\description{
The R package `sparselink` implements sparse
regression for related problems
(multi-task learning and transfer learning).
}
\details{
Use function [sparselink()] for model fitting.
Type `library(sparselink)` and then `?sparselink` or
`help("sparselink")` to open its help file.

See the vignette for further examples.
Type `vignette("sparselink")` or `browseVignettes("sparselink")`
to open the vignette.
}
\examples{
?sparselink
?predict.sparselink
?coef.sparselink

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger},
\href{https://orcid.org/0000-0003-3443-0298}{Petr N. Nazarov},
\href{https://orcid.org/0000-0003-3977-7469}{Enrico Glaab}
(2025).
"Estimating sparse regression models in multi-task learning and transfer learning through adaptive penalisation"
\emph{In preparation}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rauschenberger/sparselink}
  \item \url{https://rauschenberger.github.io/sparselink/}
  \item Report bugs at \url{https://github.com/rauschenberger/sparselink/issues}
}

}
\author{
\strong{Maintainer}: Armin Rauschenberger \email{armin.rauschenberger@uni.lu} (\href{https://orcid.org/0000-0001-6498-4801}{ORCID})

}
\keyword{documentation}
