% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{traintest}
\alias{traintest}
\title{Train and test model}
\usage{
traintest(
  y_train,
  X_train,
  y_test = NULL,
  X_test = NULL,
  family,
  alpha,
  method = c("separate", "transfer", "comb", "common"),
  type
)
}
\description{
Trains and test prediction models
}
