% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coef.sparselink}
\alias{coef.sparselink}
\title{Extract Coefficients}
\usage{
\method{coef}{sparselink}(object)
}
\arguments{
\item{object}{object of class `sparselink`}
}
\value{
Returns estimated coefficients.
The output is a list with two slots:
slot `alpha` with the estimated intercept
(vector or length \eqn{q}),
and slot `beta` with the estimated slopes
(matrix with \eqn{p} rows and \eqn{q} columns).
}
\description{
Extracts coefficients
from an object of class [sparselink].
}
\examples{
family <- "gaussian"
data <- sim.data.transfer(family=family)
#data <- sim.data.multiple(family=family)
object <- sparselink(x=data$X_train,y=data$y_train,family=family)
coef <- coef(object=object)

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger},
\href{https://orcid.org/0000-0003-3443-0298}{Petr N. Nazarov},
\href{https://orcid.org/0000-0003-3977-7469}{Enrico Glaab}
(2025).
"Estimating sparse regression models in multi-task learning and transfer learning through adaptive penalisation".
\emph{Under revision}. \url{https://hdl.handle.net/10993/63425}
}
