% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{glm.comb}
\alias{glm.comb}
\title{Sparse regression for related problems}
\usage{
glm.comb(x, y, family, alpha.init = 0.95, alpha = 1, type = "exp", nfolds = 10)
}
\arguments{
\item{x}{n x p matrix (multi-task learning) 
or list of n_k x p matrices (transfer learning)}

\item{y}{n x q matrix (multi-task learning)
or list of n_k-dimensional vectors (transfer learning)}

\item{family}{character "gaussian" or "binomial"}

\item{alpha.init}{elastic net mixing parameter for initial regressions,
default: 0.95 (lasso-like elastic net)}

\item{alpha}{elastic net mixing parameter of final regressions,
default: 1 (lasso)}

\item{type}{character}

\item{nfolds}{number of cross-validation folds}
}
\description{
Sparse regression for related problems
}
\examples{
# multi-task learning
n <- 100
p <- 50
q <- 3
family <- "gaussian"
x <- matrix(data=rnorm(n=n*p),nrow=n,ncol=p)
y <- matrix(data=rnorm(n*q),nrow=n,ncol=q)
object <- glm.comb(x=x,y=y,family=family)

}
