% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make.folds.multi}
\alias{make.folds.multi}
\title{Folds for multi-task learning}
\usage{
make.folds.multi(y, family, nfolds = 10)
}
\arguments{
\item{y}{matrix with n rows (samples) and q columns (outcomes)}

\item{family}{character "gaussian" or "binomial"}

\item{nfolds}{integer between 2 and n}
}
\description{
Folds for multi-task learning
}
\examples{

family <- "binomial"
y <- sim.data.multiple(family=family)$y_train
fold <- make.folds.multi(y=y,family=family)
table(fold)
table(y[,3],fold)

}
