% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make.folds.trans}
\alias{make.folds.trans}
\title{Folds for transfer learning}
\usage{
make.folds.trans(y, family, nfolds = 10)
}
\arguments{
\item{y}{list of q numeric vectors}

\item{family}{character "gaussian" or "binomial"}

\item{nfolds}{integer between 2 and n}
}
\description{
Folds for transfer learning
}
\examples{
family <- "binomial"
y <- sim.data.transfer(family=family)$y_train
fold <- make.folds.trans(y,family=family)

}
