% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.sparselink}
\alias{predict.sparselink}
\title{Make Predictions}
\usage{
\method{predict}{sparselink}(object, newx, weight = NULL, ...)
}
\arguments{
\item{object}{object of class `sparselink`}

\item{newx}{features:
matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}

\item{weight}{experimental argument:
numeric vector of length 2,
with the first entry for the internal weight,
and the second entry for the external weight,
overwrites the cross-validated weights}

\item{...}{(not applicable)}
}
\value{
Returns predicted values or predicted probabilities.
The output is a column vector with one entry for each sample.
}
\description{
Predicts outcome
}
\examples{
1+1

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger},
\href{https://orcid.org/0000-0003-3443-0298}{Petr N. Nazarov},
\href{https://orcid.org/0000-0003-3977-7469}{Enrico Glaab}
(2025).
"Estimating sparse regression models in multi-task learning and transfer learning through adaptive penalisation"
\emph{In preparation}.
}
