% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim.data.multiple}
\alias{sim.data.multiple}
\title{Data simulation for multi-task learning}
\usage{
sim.data.multiple(
  prob.common = 0.05,
  prob.separate = 0.05,
  q = 3,
  n0 = 100,
  n1 = 10000,
  p = 200,
  rho = 0.5,
  family = "gaussian"
)
}
\arguments{
\item{prob.common}{probability of common effect (number between 0 and 1)}

\item{prob.separate}{probability of separate effect (number between 0 and 1)}

\item{q}{number of datasets: integer}

\item{n0}{number of training samples: integer vector of length \eqn{q}}

\item{n1}{number of testing samples for all datasets: integer}

\item{p}{number of features: integer}

\item{rho}{correlation (for decreasing structure)}

\item{family}{character \code{"gaussian"} or \code{"binomial"}}
}
\value{
Returns a list with slots
\code{y_train} (\eqn{n_0 \times q} matrix),
\code{X_train}(\eqn{n_0 \times p} matrix),
\code{y_test} (\eqn{n_1 \times q} matrix),
\code{X_test} (\eqn{n_1 \times p} matrix),
and \code{beta} (\eqn{p \times q} matrix).
}
\description{
Simulates data for multi-task learning.
}
\examples{
data <- sim.data.multiple()
sapply(X=data,FUN=dim)

}
\keyword{internal}
