% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{traintest}
\alias{traintest}
\title{Train and test model}
\usage{
traintest(
  y_train,
  X_train,
  y_test = NULL,
  X_test = NULL,
  family,
  alpha,
  method = c("glm.empty", "glm.separate", "sparselink"),
  alpha.init,
  type,
  trial = FALSE
)
}
\arguments{
\item{y_train}{target of training samples:
\eqn{n \times q} matrix (multi-task learning)
or list of \eqn{q} vectors of length \eqn{n_1,\ldots,n_q} (transfer learning)}

\item{X_train}{features of training samples:
\eqn{n \times p} matrix (multi-task learning)
or list of \eqn{q} matrices of dimensions
\eqn{n_1 \times p,\ldots,n_q \times p} (transfer learning)}

\item{y_test}{target of testing samples:
\eqn{m \times p} matrix (multi-task learning)
or list of \eqn{q} vectors of length \eqn{m_1,\ldots,m_q} (transfer learning)}

\item{X_test}{features of testing samples:
\eqn{m \times p} matrix (multi-task learning) or
list of \eqn{q} matrices of dimensions
\eqn{m_1 \times p,\ldots,m_q \times p} (transfer learning)}

\item{family}{character \code{"gaussian"} or \code{"binomial"}}

\item{alpha}{elastic net mixing parameter}

\item{method}{character vector}

\item{alpha.init}{elastic net mixing parameter for initial regressions}

\item{type}{character}

\item{trial}{see sparselink}
}
\description{
Trains and tests prediction models
}
\examples{
#--- multi-task learning ---
\dontrun{
family <- "gaussian"
data <- sim.data.multiple(family=family)
result <- traintest(data$y_train,data$X_train,family=family,alpha=1,method="glm.separate",alpha.init=0.95,type="exp")
}

#--- transfer learning ---
\dontrun{
family <- "gaussian"
data <- sim.data.transfer(family=family)
result <- traintest(data$y_train,data$X_train,family=family,alpha=1,alpha.init=0.95,type="exp")
}

}
\keyword{internal}
