% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{traintest}
\alias{traintest}
\title{Train and test model}
\usage{
traintest(
  y_train,
  X_train,
  y_test = NULL,
  X_test = NULL,
  family,
  alpha,
  method = c("separate", "transfer", "comb", "common"),
  alpha.init,
  type
)
}
\arguments{
\item{y_train}{target of training samples: vector of length n}

\item{X_train}{features of training samples: n x p matrix}

\item{y_test}{target of testing samples: vector of length m}

\item{X_test}{features of testing samples m x p matrix}

\item{family}{character "gaussian" or "binomial"}

\item{alpha}{elastic net mixing paramter}

\item{method}{character vector}

\item{alpha.init}{elastic net mixing parameter for initial regressions}

\item{type}{character}
}
\description{
Trains and test prediction models
}
