% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{traintest}
\alias{traintest}
\title{Train and test model}
\usage{
traintest(
  y_train,
  X_train,
  y_test = NULL,
  X_test = NULL,
  family,
  alpha,
  method = c("glm.separate", "glm.glmtrans", "sparselink", "glm.common"),
  alpha.init,
  type,
  trial = FALSE
)
}
\arguments{
\item{y_train}{target of training samples: vector of length \eqn{n}}

\item{X_train}{features of training samples: \eqn{n \times p} matrix}

\item{y_test}{target of testing samples: vector of length \eqn{m}}

\item{X_test}{features of testing samples \eqn{m \times p} matrix}

\item{family}{character \code{"gaussian"} or \code{"binomial"}}

\item{alpha}{elastic net mixing parameter}

\item{method}{character vector}

\item{alpha.init}{elastic net mixing parameter for initial regressions}

\item{type}{character}

\item{trial}{see sparselink}
}
\description{
Trains and tests prediction models
}
